+++ tests/unit/deploy/test_swagger.py
@@ -379,7 +379,31 @@ def test_can_add_api_key(sample_app, swagger_gen):
 
 def test_can_use_authorizer_object(sample_app, swagger_gen):
     authorizer = CustomAuthorizer(
-        'MyAuth', authorizer_uri='auth-uri', header='Authorization')
+        'MyAuth', authorizer_uri='auth-uri', header='Authorization'
+    )
+
+    @sample_app.route('/auth', authorizer=authorizer)
+    def auth():
+        return {'foo': 'bar'}
+
+    doc = swagger_gen.generate_swagger(sample_app)
+    single_method = doc['paths']['/auth']['get']
+    assert single_method.get('security') == [{'MyAuth': []}]
+    security_definitions = doc['securityDefinitions']
+    assert 'MyAuth' in security_definitions
+    my_auth = security_definitions['MyAuth']
+    # authorizerCredentials should not be in this dict because it's None.
+    assert my_auth['x-amazon-apigateway-authorizer'] == {
+        'authorizerUri': 'auth-uri',
+        'type': 'token',
+        'authorizerResultTtlInSeconds': 300,
+    }
+
+
+def test_can_use_authorizer_object_with_role_arn(sample_app, swagger_gen):
+    authorizer = CustomAuthorizer(
+        'MyAuth', authorizer_uri='auth-uri', header='Authorization',
+        invoke_role_arn='role-arn')
 
     @sample_app.route('/auth', authorizer=authorizer)
     def auth():
@@ -398,7 +422,8 @@ def test_can_use_authorizer_object(sample_app, swagger_gen):
         'x-amazon-apigateway-authorizer': {
             'authorizerUri': 'auth-uri',
             'type': 'token',
-            'authorizerResultTtlInSeconds': 300
+            'authorizerResultTtlInSeconds': 300,
+            'authorizerCredentials': 'role-arn'
         }
     }